+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-05-20
+
+### Internal Changes
+- Feat: Add Inertia version 2 DeferredProp feature
+- Feat: Add async support for partial loading
+- Examples: Add code example for Inertia version 2 support features
+
 ## [1.0.6] - 2025-05-13
 
 - Add config option to pass extra context to the template via the `extra_template_context` option in `InertiaConfig`